// hr-frontend/src/lib/api/generated/model/deletedEmployeeSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface DeletedEmployeeSummary {
  employeeId?: number;
  userId?: number;
  employeeNumber?: string;
  displayName?: string;
  deletionReason?: string;
  deletedAt?: string;
  deletedByUserId?: number;
}